Add network-util at common
